project/requirements.txt
Django==1.9.5
django-filter==0.13.0
djangorestframework==3.3.3
Jinja2==2.8
MarkupSafe==0.23
requests==2.9.1
uWSGI==2.0.12
wheel==0.24.0
